pyproject.toml
[project]
name = "kimchi-premium-strategy-implementation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.13",
    "apscheduler>=3.11.0",
    "boto3>=1.38.41",
    "celery>=5.5.3",
    "fastapi>=0.115.13",
    "pybit>=5.11.0",
    "pyjwt>=2.10.1",
    "pytest-asyncio>=1.0.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "redis>=6.2.0",
    "uvicorn>=0.34.3",
]
